pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "flask>=3.1.0",
    "python-dotenv==1.0.1",
]
